Adjusting order of XML elements to match the UPnP Device Architecture 1.1 XML schema.
cling implements UPnP 1.0.

Whilst the UPnP Device Architecture 1.0 document doesn't provide a full XML schema, the
1.1 document does so. It's available from:
http://upnp.org/sdcps-and-certification/standards/device-architecture-documents/

In Appendix B1 it gives the full schema for a UPnP device, and in Appendix B2 for
a UPnP service.

The order of XML elements is important in this schemas, since nearly all of the
parent elements are defined as sequences with fixed order.

This change amends cling to match the order given in these two appendices.
Since cling is implementing UPnP 1.0, it shouldn't really matter, but we know
of at least one UPnP stack which insists that the order of elements is the same
as given in this 1.1 schema, even if the UPnP version in use is 1.0.
(This stack is somehow based upon GUPnP and Qt, but we don't have full visibility
of its internals - we just want to make cling work with it!)

Specifically this change:
* Moves the UDN into the right place in the device XML.
* Moves the SCPDURL into the right place in the device XML.
* Moves the 'relatedStateVariable' into the right place in the service XML.