Refactors to minimize the use of URI.create.
In the current implementation of Namespace there's a number of
calls to URI.create() which result in an unnecessary encoding and decoding
of URI/URL objects.

We have refactored and enhanced the Namespace class so that it will return Uri paths
as strings (getDescriptorUriPath(Device device),getEventCallbackUriPath(Service service))

We've done this to avoid the use of URI.create, and instead use a more explicit URL constructor
that avoids unnecessary rencoding operations.

Wherever you needed to use Location before you had to pass paths as URI objects which had to
be created using Namespace methods, these methods would unnecessarily build URI objects the
slow way (parsing, re-encoding)

RemoteGENASubscription and SendingNotification were the only classes affected by the api changes.